Change authorizeGroupMember() to accept an array of authorized users
Now you can pass an array of user ids that are authorized to access the group's resources. These users don't have to be masters of the group even if requireMaster is set to true. Useful for situations where the user has to be able to do self-targeted actions which would require master rights if done to other users.